refactor: 이미지 검증 ImageValidator로 통합
