+++ src/main/java/com/ururulab/ururu/global/config/SecurityConfig.java
@@ -43,8 +43,6 @@ public class SecurityConfig {
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers("/api/auth/**").permitAll()
                         .requestMatchers("/api/public/**").permitAll()
-                        .requestMatchers("/products/**").permitAll() // 추후 삭제
-                        .requestMatchers("/groupbuy/**").permitAll() // 추후 삭제
                         .requestMatchers("/health", "/actuator/health").permitAll()
                         .requestMatchers("/swagger-ui/**", "/v3/api-docs/**").permitAll()
                         .requestMatchers("/h2-console/**").permitAll()
