+++ src/main/java/com/ururulab/ururu/product/service/validation/ProductValidator.java
@@ -5,6 +5,7 @@ import com.ururulab.ururu.global.exception.BusinessException;
 import com.ururulab.ururu.global.exception.error.ErrorCode;
 import com.ururulab.ururu.image.domain.ImageFormat;
 import com.ururulab.ururu.image.exception.InvalidImageFormatException;
+import com.ururulab.ururu.image.validation.ImageValidator;
 import com.ururulab.ururu.product.controller.dto.request.ProductOptionRequest;
 import com.ururulab.ururu.product.domain.entity.Category;
 import com.ururulab.ururu.product.service.CategoryCacheService;
@@ -22,6 +23,7 @@ import java.util.Optional;
 public class ProductValidator {
 
     private final CategoryCacheService categoryCacheService;
+    private final ImageValidator imageValidator;
 
     /**
      * 카테고리 유효성 검증
@@ -45,7 +47,7 @@ public class ProductValidator {
 
     public void validateOptionImagePair(List<ProductOptionRequest> options, List<MultipartFile> images) {
         validateOptionImageCount(options, images);
-        validateAllImages(images);
+        imageValidator.validateAllImages(images);
     }
 
     /**
@@ -58,86 +60,4 @@ public class ProductValidator {
                     options.size(), images.size());
         }
     }
-
-    /**
-     * 상품 옵션 이미지 검증
-     */
-    public void validateImage(MultipartFile image) {
-        long start = System.currentTimeMillis();
-
-        if (image == null || image.isEmpty()) {
-            return;
-        }
-
-        ImageFormat extFmt = parseExtension(image);
-        ImageFormat mimeFmt = parseMimeType(image);
-        ensureMatchingFormats(extFmt, mimeFmt, image);
-
-        long end = System.currentTimeMillis();
-        log.info("Image validation took: {}ms for file: {}", end - start, image.getOriginalFilename());
-    }
-
-    public void validateAllImages(List<MultipartFile> optionImages) {
-        if (optionImages == null || optionImages.isEmpty()) {
-            return;
-        }
-
-        for (int i = 0; i < optionImages.size(); i++) {
-            MultipartFile imageFile = optionImages.get(i);
-
-            if (imageFile == null || imageFile.isEmpty()) {
-                continue;
-            }
-
-            try {
-                validateImage(imageFile);  // 기존 메서드 재사용
-            } catch (Exception e) {
-                throw new IllegalArgumentException(
-                        String.format("옵션 %d번째 이미지가 유효하지 않습니다: %s", i + 1, e.getMessage()), e
-                );
-            }
-        }
-    }
-
-    private ImageFormat parseExtension(MultipartFile file) {
-        String filename = Optional.ofNullable(file.getOriginalFilename())
-                .filter(n -> n.contains("."))
-                .orElseThrow(() ->
-                        new InvalidImageFormatException("파일명이 없거나 확장자를 찾을 수 없습니다.")
-                );
-        int idx = filename.lastIndexOf('.');
-        if (idx == filename.length() - 1) {
-            throw new InvalidImageFormatException("파일명이 마침표로 끝납니다: " + filename);
-        }
-        String ext = filename.substring(idx + 1).toLowerCase();
-        return ImageFormat.fromExtension(ext)
-                .orElseThrow(() -> new InvalidImageFormatException("지원하지 않는 확장자: " + ext));
-    }
-
-    private ImageFormat parseMimeType(MultipartFile file) {
-        String mime = Optional.ofNullable(file.getContentType())
-                .orElseThrow(() ->
-                        new InvalidImageFormatException("MIME 타입을 확인할 수 없습니다.")
-                );
-        return ImageFormat.fromMimeType(mime)
-                .orElseThrow(() -> new InvalidImageFormatException("지원하지 않는 MIME 타입: " + mime));
-    }
-
-    private void ensureMatchingFormats(
-            ImageFormat extFmt,
-            ImageFormat mimeFmt,
-            MultipartFile file
-    ) {
-        if (!extFmt.getMimeType().equals(mimeFmt.getMimeType())) {
-            throw new InvalidImageFormatException(
-                    String.format(
-                            "확장자(%s)와 MIME(%s)이 일치하지 않습니다: file=%s",
-                            extFmt.getExtension(),
-                            mimeFmt.getMimeType(),
-                            file.getOriginalFilename()
-                    )
-            );
-        }
-    }
-
 }
